treeproject-main/requirements.txt
# Core Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# ML & Computer Vision
ultralytics==8.2.72
torch==2.2.2
torchvision==0.17.2
opencv-python==4.10.0.84
Pillow==10.4.0

# Geospatial
rasterio==1.3.10
pyproj==3.5.0
shapely==2.0.5
pygeoops==0.4.0
geopandas==0.13.2
fiona==1.9.6

# Scientific Computing
numpy==1.24.4
scipy==1.10.1
pandas==2.0.3

# Utilities
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3

# Optional: GDAL for TIFF processing
# Note: GDAL installation can be complex, see INSTALL_GDAL.md
# GDAL==3.4.3

# Development
gunicorn==21.2.0  # Production WSGI server

